#5384 [Manager] Add threshold of coverage (#5387)

+++ ohara-manager/scripts/copyJars.js
@@ -16,6 +16,7 @@
 
 const fs = require('fs');
 const execa = require('execa');
+const package = require('../package.json');
 
 const srcBase = '../ohara-it/build/libs';
 const distBase = './client/cypress/fixtures/jars';
@@ -23,28 +24,32 @@ const files = [
   'ohara-it-source.jar',
   'ohara-it-sink.jar',
   'ohara-it-stream.jar',
-  'ohara-it-0.11.0-SNAPSHOT.jar',
+  `ohara-it-${package.version}.jar`,
 ];
 
-/* eslint-disable no-process-exit, no-console */
-const buildJars = execa('./gradlew', ['ohara-it:jar'], {
-  cwd: '..',
-  stdio: 'inherit',
-});
+const filesExist = files.every((file) => fs.existsSync(`${distBase}/${file}`));
 
-buildJars
-  .then((value) => {
-    console.log(`build jars output: ${value.stdout}`);
+if (!filesExist) {
+  /* eslint-disable no-process-exit, no-console */
+  (async function () {
+    try {
+      const value = await execa('./gradlew', ['ohara-it:jar'], {
+        cwd: '..',
+        stdio: 'inherit',
+      });
 
-    if (!fs.existsSync(`${distBase}`)) {
-      fs.mkdirSync(`${distBase}`, { recursive: true });
+      console.log(`build jars output: ${value.stdout}`);
+
+      if (!fs.existsSync(`${distBase}`)) {
+        fs.mkdirSync(`${distBase}`, { recursive: true });
+      }
+
+      files.forEach((file) => {
+        fs.copyFileSync(`${srcBase}/${file}`, `${distBase}/${file}`);
+      });
+    } catch (err) {
+      console.log(err.message);
+      process.exit(1);
     }
-
-    files.forEach((file) => {
-      fs.copyFileSync(`${srcBase}/${file}`, `${distBase}/${file}`);
-    });
-  })
-  .catch((err) => {
-    console.log(err.message);
-    process.exit(1);
-  });
+  })();
+}
